requirements.txt
# LLM
groq>=0.9.0
ollama>=0.3.0

# AWS
boto3>=1.34.0
aws-cdk-lib>=2.140.0
constructs>=10.3.0

# Data
yfinance>=0.2.40
fredapi>=0.5.2
newsapi-python>=0.2.7
praw>=7.7.1
entsoe-py>=0.6.9
pandas>=2.2.0
numpy>=1.26.0
ta>=0.11.0

# Web / API
requests>=2.31.0

# Validation
pydantic>=2.7.0
python-dotenv>=1.0.0

# Dashboard
streamlit>=1.35.0
plotly>=5.22.0
matplotlib>=3.8.0

# Testing
pytest>=8.2.0
pytest-mock>=3.14.0
moto[s3,dynamodb,sqs,secretsmanager]>=5.0.0

# Linting
ruff>=0.4.0
